requirements.txt
flask==1.0.0
requests==2.18.4
gunicorn==19.9.0
python-dotenv==0.12.0
nltk==3.4.5
rake-nltk==1.0.4
Pillow==7.1.1
pyocr==0.7.2
textblob==0.15.3
numpy==1.18.2
networkx==2.4
pytesseract